requirements.txt
tavily-python>=0.5.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pyyaml>=6.0.2
flask>=3.0.0
